Adding the ability to set whether a user is active or not rather than deleting them, stops them from being able to log in to the system.
