+++ packages/server/src/api/controllers/user.js
@@ -42,6 +42,10 @@ exports.create = async function(ctx) {
     password: hashedPassword,
     tableId: ViewNames.USERS,
   }
+  // add the active status to a user if its not provided
+  if (typeof user.active !== "boolean") {
+    user.active = true
+  }
 
   try {
     const response = await db.post(user)
@@ -64,13 +68,21 @@ exports.create = async function(ctx) {
 exports.update = async function(ctx) {
   const db = new CouchDB(ctx.user.appId)
   const user = ctx.request.body
+  let dbUser
+  // get user incase password removed
+  if (user._id) {
+    dbUser = await db.get(user._id)
+  }
   if (user.password) {
     user.password = await bcrypt.hash(user.password)
   } else {
     delete user.password
   }
 
-  const response = await db.put(user)
+  const response = await db.put({
+    password: dbUser.password,
+    ...user,
+  })
   user._rev = response.rev
 
   ctx.status = 200
