+++ packages/server/src/constants/index.js
@@ -1,44 +1,5 @@
 const { BUILTIN_ROLE_IDS } = require("../utilities/security/roles")
 
-const AuthTypes = {
-  APP: "app",
-  BUILDER: "builder",
-  EXTERNAL: "external",
-}
-
-const USERS_TABLE_SCHEMA = {
-  _id: "ta_users",
-  type: "table",
-  views: {},
-  name: "Users",
-  schema: {
-    email: {
-      type: "string",
-      constraints: {
-        type: "string",
-        email: true,
-        length: {
-          maximum: "",
-        },
-        presence: true,
-      },
-      fieldName: "email",
-      name: "email",
-    },
-    roleId: {
-      fieldName: "roleId",
-      name: "roleId",
-      type: "options",
-      constraints: {
-        type: "string",
-        presence: false,
-        inclusion: Object.values(BUILTIN_ROLE_IDS),
-      },
-    },
-  },
-  primaryDisplay: "email",
-}
-
 exports.FieldTypes = {
   STRING: "string",
   LONGFORM: "longform",
@@ -51,6 +12,54 @@ exports.FieldTypes = {
   AUTO: "auto",
 }
 
+exports.AuthTypes = {
+  APP: "app",
+  BUILDER: "builder",
+  EXTERNAL: "external",
+}
+
+exports.USERS_TABLE_SCHEMA = {
+  _id: "ta_users",
+  type: "table",
+  views: {},
+  name: "Users",
+  schema: {
+    email: {
+      type: exports.FieldTypes.STRING,
+      constraints: {
+        type: exports.FieldTypes.STRING,
+        email: true,
+        length: {
+          maximum: "",
+        },
+        presence: true,
+      },
+      fieldName: "email",
+      name: "email",
+    },
+    roleId: {
+      fieldName: "roleId",
+      name: "roleId",
+      type: exports.FieldTypes.OPTIONS,
+      constraints: {
+        type: exports.FieldTypes.STRING,
+        presence: false,
+        inclusion: Object.values(BUILTIN_ROLE_IDS),
+      },
+    },
+    active: {
+      fieldName: "active",
+      name: "active",
+      type: exports.FieldTypes.BOOLEAN,
+      constraints: {
+        type: exports.FieldTypes.BOOLEAN,
+        presence: false,
+      },
+    },
+  },
+  primaryDisplay: "email",
+}
+
 exports.AutoFieldSubTypes = {
   CREATED_BY: "createdBy",
   CREATED_AT: "createdAt",
@@ -59,8 +68,6 @@ exports.AutoFieldSubTypes = {
   AUTO_ID: "autoID",
 }
 
-exports.AuthTypes = AuthTypes
-exports.USERS_TABLE_SCHEMA = USERS_TABLE_SCHEMA
 exports.BUILDER_CONFIG_DB = "builder-config-db"
 exports.HOSTING_DOC = "hosting-doc"
 exports.OBJ_STORE_DIRECTORY = "/app-assets/assets"
